Added eye shader file
In the project for which I made this, the shaders and scripts were in the same folder, hence I place it here. Feel free to move it somewhere else. @leon have fun reading it; it's not that hard, but it's just a lot of repetition of declarations of variables etc... just weird.